+++ src/routines/copy.cc
@@ -90,7 +90,7 @@ xkblas_t::copy_tile_async(
     new (dep) task_dep_info_t(AC);
 
     task_dev_info_t * dev = TASK_DEV_INFO(task);
-    constexpr size_t ocr_access = 1;
+    constexpr int ocr_access = 1;
     new (dev) task_dev_info_t(device_global_id, ocr_access);
 
     # if XKRT_SUPPORT_DEBUG
@@ -141,24 +141,24 @@ xkblas_t::copy_async(
 
     // get tile size
     xkblas_t * context = xkblas_get();
-    size_t ts = context->conf.kernels[COPY].tile;
+    int ts = context->conf.kernels[COPY].tile;
     if (ts == 0)
     {
         int args[1] = { (int) n };
         xkblas_routine_auto_tile(COPY, args, &ts);
     }
-    const size_t nt = NUM_OF_TILES(n, ts);
+    const int nt = NUM_OF_TILES(n, ts);
 
     // get number of gpus
-    const size_t ngpus = context->runtime.get_ndevices() - 1;
+    const int ngpus = context->runtime.get_ndevices() - 1;
     distribution_t d;
     distribution1D_init(&d, XKRT_DISTRIBUTION_TYPE_CYCLIC1D, ngpus, n, ts);
 
     // spawn tiles
-    for (size_t tn = 0 ; tn < nt ; ++tn)
+    for (int tn = 0 ; tn < nt ; ++tn)
     {
-        size_t offset = tn*ts;
-        size_t bs     = (tn == nt-1) ? (n - offset) : ts;
+        int offset = tn*ts;
+        int bs     = (tn == nt-1) ? (n - offset) : ts;
         device_global_id_t device_global_id = distribution1D_get(&d, tn);
         this->copy_tile_async<P>(bs, x + offset, incx, y + offset, incy, device_global_id);
     }
