+++ src/routines/dot.cc
@@ -44,7 +44,7 @@ TYPED
 struct args_t
 {
     args_t(
-        size_t n,
+        int n,
         int incx,
         int incy,
         TYPE * r
@@ -57,7 +57,7 @@ struct args_t
 
     ~args_t() {}
 
-    const size_t n;
+    const int n;
     const int incx;
     const int incy;
     TYPE * r;
@@ -88,7 +88,7 @@ xkblas_t::dot_tile_async(
     new (dep) task_dep_info_t(AC);
 
     task_dev_info_t * dev = TASK_DEV_INFO(task);
-    constexpr size_t ocr_access = 0;
+    constexpr int ocr_access = 0;
     new (dev) task_dev_info_t(device_global_id, ocr_access);
 
     args_t<P> * args = (args_t<P> *) TASK_ARGS(task, task_size);
@@ -128,13 +128,13 @@ xkblas_t::dot_async(
 
     // get tile size
     xkblas_t * xkblas = xkblas_get();
-    size_t ts = xkblas->conf.kernels[DOT].tile;
+    int ts = xkblas->conf.kernels[DOT].tile;
     if (ts == 0)
     {
         int args[1] = { n };
         xkblas_routine_auto_tile(DOT, args, &ts);
     }
-    const size_t nt = NUM_OF_TILES(n, ts);
+    const int nt = NUM_OF_TILES(n, ts);
 
     if (nt == 1)
     {
@@ -152,9 +152,9 @@ xkblas_t::dot_async(
         distribution1D_init(&d, XKRT_DISTRIBUTION_TYPE_CYCLIC1D, ngpus, n, ts);
 
         // spawn tiles
-        for (size_t tn = 0 ; tn < nt ; ++tn)
+        for (int tn = 0 ; tn < nt ; ++tn)
         {
-            size_t bs = (tn == nt-1) ? (n - tn*ts) : ts;
+            int bs = (tn == nt-1) ? (n - tn*ts) : ts;
             device_global_id_t device_global_id = distribution1D_get(&d, tn);
             this->dot_tile_async<P>(bs, x + tn*ts*incx, incx, y + tn*ts*incy, incy, temp_r, temp_r + tn, device_global_id);
         }
@@ -168,7 +168,7 @@ xkblas_t::dot_async(
             [=] (runtime_t * runtime, device_t * device, task_t * task) {
                 (void) runtime; (void) device; (void) task;
                 *r = 0;
-                for (size_t i = 0 ; i < nt ; ++i)
+                for (int i = 0 ; i < nt ; ++i)
                     *r += temp_r[i];
                 free(temp_r);
             }
